refactor: Header, Footer widget으로 위치 변경
