+++ src/widgets/Footer.tsx
